# src/models/product.js
class Product {
    constructor(id, name, price, stock, description, imageUrl) {
      this.id = id;
      this.name = name;
      this.price = price;
      this.stock = stock;
      this.description = description;
      this.imageUrl = imageUrl;
    }
  }
  
  module.exports = Product;
  